test(item-sliding): get interactive test working locally
moves open and close functions to utils
