+++ core/src/components/item-sliding/test/standalone/e2e.ts
@@ -1,5 +1,7 @@
 import { newE2EPage } from '@stencil/core/testing';
 
+import { closeItemSliding, openItemSliding } from '../test.utils';
+
 test('item-sliding: standalone', async () => {
   const page = await newE2EPage({
     url: '/src/components/item-sliding/test/standalone?ionic:_testing=true'
@@ -9,50 +11,21 @@ test('item-sliding: standalone', async () => {
   compares.push(await page.compareScreenshot());
 
   // Pass sliding item with start icons in option
-  await openItem('#startItem', page);
+  await openItemSliding('#startItem', page);
   compares.push(await page.compareScreenshot(`start icons open`));
-  await closeItem(page);
+  await closeItemSliding(page);
 
   // Pass sliding item with top icons in option
-  await openItem('#topItem', page);
+  await openItemSliding('#topItem', page);
   compares.push(await page.compareScreenshot(`top icons open`));
-  await closeItem(page);
+  await closeItemSliding(page);
 
   // Pass sliding item with anchor option
-  await openItem('#anchorItem', page);
+  await openItemSliding('#anchorItem', page);
   compares.push(await page.compareScreenshot(`anchor option`));
-  await closeItem(page);
+  await closeItemSliding(page);
 
   for (const compare of compares) {
     expect(compare).toMatchScreenshot();
   }
 });
-
-// Opens a sliding item by simulating a drag event
-async function openItem(id: string, page: any) {
-  try {
-    const slidingItem = await page.$(id);
-
-    // Simulate a drag
-    const boundingBox = await slidingItem.boundingBox();
-    const centerX = parseFloat(boundingBox.x + boundingBox.width / 2);
-    const centerY = parseFloat(boundingBox.y + boundingBox.height / 2);
-
-    await page.mouse.move(centerX, centerY);
-    await page.mouse.down();
-    await page.mouse.move(centerX / 2, centerY);
-    await page.mouse.move(0, centerY);
-    await page.mouse.up();
-  } catch (err) {
-    throw err;
-  }
-}
-
-// Close a sliding item after taking a screenshot
-// to allow other sliding items to open
-async function closeItem(page: any) {
-  await page.mouse.move(0, 0);
-  await page.mouse.down();
-  await page.mouse.up();
-  await page.waitFor(1000);
-}
